// src/clarity-angular/wigfrid/enum/Key.d.ts
/**
 * Enumeration with key values.
 *
 * This enumeration is useful when handling <b>keyDown</b> events.
 */
export declare enum Key {
    /** The backspace key. */
    Back = 8,
    /** The tab key. */
    Tab = 9,
    /** The enter key. */
    Enter = 13,
    /** The escape key. */
    Escape = 27,
    /** The space key. */
    Space = 32,
    /** The page up key. */
    PageUp = 33,
    /** The page down key. */
    PageDown = 34,
    /** The end key. */
    End = 35,
    /** The home key. */
    Home = 36,
    /** The left arrow key. */
    Left = 37,
    /** The up arrow key. */
    Up = 38,
    /** The right arrow key. */
    Right = 39,
    /** The down arrow key. */
    Down = 40,
    /** The delete key. */
    Delete = 46,
    /** The F1 key. */
    F1 = 112,
    /** The F2 key. */
    F2 = 113,
    /** The F3 key. */
    F3 = 114,
    /** The F4 key. */
    F4 = 115,
    /** The F5 key. */
    F5 = 116,
    /** The F6 key. */
    F6 = 117,
    /** The F7 key. */
    F7 = 118,
    /** The F8 key. */
    F8 = 119,
    /** The F9 key. */
    F9 = 120,
    /** The F10 key. */
    F10 = 121,
    /** The F11 key. */
    F11 = 122,
    /** The F12 key. */
    F12 = 123,
}